pkg/loop/internal: move interfaces to types.go
