+++ pkg/loop/internal/reporting.go
@@ -12,13 +12,7 @@ import (
 	libocr "github.com/smartcontractkit/libocr/offchainreporting2plus/types"
 
 	"github.com/smartcontractkit/chainlink-relay/pkg/loop/internal/pb"
-	"github.com/smartcontractkit/chainlink-relay/pkg/types"
 )
-
-type ReportingPluginFactory interface {
-	types.Service
-	libocr.ReportingPluginFactory
-}
 
 type reportingPluginFactoryClient struct {
 	*brokerExt
